+++ plugin/src/main/kotlin/de/fayard/BuildSrcVersionsExtension.kt
@@ -1,13 +1,13 @@
 package de.fayard
 
+import com.github.benmanes.gradle.versions.updates.DependencyUpdatesTask
+import com.github.benmanes.gradle.versions.updates.resolutionstrategy.ComponentFilter
+
+
 interface BuildSrcVersionsExtension {
-    var useFdqnFor: MutableList<String>
+    fun useFqdnFor(vararg dependencyName: String)
 
-    fun useFdqnFor(vararg dependencyName: String)
-
-    var rejectedVersionKeywords: MutableList<String>
-
-    fun rejectedVersionKeywords(vararg keyword: String)
+    fun rejectVersionIf(filter: ComponentFilter)
 
     var renameLibs : String
 
@@ -19,28 +19,53 @@ interface BuildSrcVersionsExtension {
 
     var versionsOnlyFile : String?
 
+
+    @Deprecated("Deprecated, see #64", ReplaceWith("useFqdnFor()"))
+    var useFdqnFor: MutableList<String>
+
+    @Deprecated("Deprecated, see #64", ReplaceWith("useFqdnFor(dependencyName)"))
+    fun useFdqnFor(vararg dependencyName: String)
+
+
+    @Deprecated("Remove or use rejectVersionIf { ... }", replaceWith = ReplaceWith(""))
+    var rejectedVersionKeywords: MutableList<String>
+
+    @Deprecated("Remove or use rejectVersionIf { selection -> ... }", replaceWith = ReplaceWith(""))
+    fun rejectedVersionKeywords(vararg keyword: String)
 }
 
-open class BuildSrcVersionsExtensionImpl(
-    override var useFdqnFor: MutableList<String> = mutableListOf(),
+internal open class BuildSrcVersionsExtensionImpl(
     override var renameLibs: String = PluginConfig.DEFAULT_LIBS,
     override var renameVersions: String = PluginConfig.DEFAULT_VERSIONS,
     override var indent: String = PluginConfig.DEFAULT_INDENT,
     override var versionsOnlyMode: VersionsOnlyMode? = null,
-    override var versionsOnlyFile: String? = null,
-    override var rejectedVersionKeywords: MutableList<String> = PluginConfig.DEFAULT_REJECTED_KEYWORDS
+    override var versionsOnlyFile: String? = null
 ) : BuildSrcVersionsExtension {
+    var useFqqnFor: List<String> = emptyList()
 
+    // Use @Transient for fields that should not be present in toString()
+    override fun toString() : String = PluginConfig.extensionAdapter.toJson(this)
+
+    @Transient lateinit var upstream: DependencyUpdatesTask
+
+    override fun rejectVersionIf(filter: ComponentFilter) {
+        upstream.rejectVersionIf(filter)
+    }
+
+    override fun useFqdnFor(vararg dependencyName: String) {
+        useFqqnFor = dependencyName.toList()
+    }
 
     override fun rejectedVersionKeywords(vararg keyword: String) {
-        rejectedVersionKeywords = keyword.toMutableList()
+        println("Warning: rejectedVersionKeywords is deprecated, see ${PluginConfig.issue53PluginConfiguration}")
     }
 
     override fun useFdqnFor(vararg dependencyName: String) {
-        useFdqnFor = dependencyName.toMutableList()
+        println("Warning: useFdqnFor is deprecated, use useFqqnFor() instead. ${PluginConfig.issue53PluginConfiguration}")
+        useFqdnFor(*dependencyName)
     }
 
-    override fun toString() : String =
-        PluginConfig.extensionAdapter.toJson(this)
+    @Transient override var useFdqnFor: MutableList<String> = mutableListOf()
+    @Transient override var rejectedVersionKeywords: MutableList<String> = mutableListOf()
 
 }