+++ plugin/src/main/kotlin/de/fayard/BuildSrcVersionsPlugin.kt
@@ -1,18 +1,21 @@
 package de.fayard
 
 import com.github.benmanes.gradle.versions.updates.DependencyUpdatesTask
+import com.github.benmanes.gradle.versions.updates.resolutionstrategy.ComponentSelectionWithCurrent
+import de.fayard.PluginConfig.isNonStable
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.create
-import org.gradle.kotlin.dsl.getByType
 
 open class BuildSrcVersionsPlugin : Plugin<Project> {
 
     override fun apply(project: Project) = project.run {
 
-        configureBenManesVersions()
-
-        extensions.create(BuildSrcVersionsExtension::class, PluginConfig.EXTENSION_NAME, BuildSrcVersionsExtensionImpl::class)
+        val extension = extensions.create(BuildSrcVersionsExtension::class, PluginConfig.EXTENSION_NAME, BuildSrcVersionsExtensionImpl::class)
+        (extension as BuildSrcVersionsExtensionImpl).upstream = configureBenManesVersions()
+        extension.rejectVersionIf {
+            isNonStable(candidate.version)
+        }
 
         tasks.create("buildSrcVersions", BuildSrcVersionsTask::class) {
             group = "Help"
@@ -20,32 +23,12 @@ open class BuildSrcVersionsPlugin : Plugin<Project> {
             dependsOn(":dependencyUpdates")
             outputs.upToDateWhen { false }
         }
-
         Unit
     }
 
-    fun Project.configureBenManesVersions(): DependencyUpdatesTask {
-        val rejectedKeywordsRegexps: List<Regex> by lazy {
-            project.extensions.getByType<BuildSrcVersionsExtension>().rejectedVersionKeywords
-                .map { qualifier -> Regex("(?i).*[.-]$qualifier[.\\d-]*") }
+    fun Project.configureBenManesVersions(): DependencyUpdatesTask =
+        tasks.maybeCreate("dependencyUpdates", DependencyUpdatesTask::class.java).also { task: DependencyUpdatesTask ->
+            task.checkForGradleUpdate = true
+            task.outputFormatter = "json"
         }
-
-        val benManesVersions: DependencyUpdatesTask =
-            tasks.maybeCreate("dependencyUpdates", DependencyUpdatesTask::class.java)
-
-        benManesVersions.outputFormatter = "json"
-        benManesVersions.checkForGradleUpdate = true
-        benManesVersions.resolutionStrategy {
-
-            componentSelection {
-                all {
-                    if (rejectedKeywordsRegexps.any { it.matches(candidate.version) }) {
-                        reject("Release candidate")
-                    }
-                }
-            }
-
-        }
-        return benManesVersions
-    }
 }